Introduced Plugin System
This plugin system fundamentally changes the original code base by moving all of the Unreal types to the new plugin core library, leaving the main library 'UnrealUAssetConverter' bare, but necessary.

Going forward any functions that may need to be accessible to plugins through the UAssetConverter must be added to IAssetConverter found in the UConvertPlugin project.