Implemented Reading Export Properties and Code Cleanup
The Export Properties can now be exported using a reference to an export object. Maybe a todo would be to allow specific objects to be requested by name/path.

Updated all of the GetXMap functions to preserve previous call's position in the stream to prevent seeking errors.

Updated the example program to display which plugins have been loaded and outputs the export types and properties if it is able to find them.

Changed the plugin type to use a "TagData" and "TagValue" system instead of the `pre` bool. This makes explicit usage much more obvious and improves readability.

Updated the ByteProperty example plugin to use the correct stream. Changed from `GetAssetStream()` to `GetExportStream()`.

Changed the LoadAllPlugins function to return a list of types instead of instances of IConverterPlugin, this allows instances to be spawned on demand if needed.